+++ src/main/java/teamfusion/bettas/block/DriedLeavesBlock.java
@@ -1,0 +1,101 @@
+package teamfusion.bettas.block;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.context.BlockPlaceContext;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.BushBlock;
+import net.minecraft.world.level.block.SimpleWaterloggedBlock;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.level.material.Fluids;
+import net.minecraft.world.phys.shapes.CollisionContext;
+import net.minecraft.world.phys.shapes.VoxelShape;
+
+import javax.annotation.Nullable;
+
+public class DriedLeavesBlock extends BushBlock implements SimpleWaterloggedBlock {
+    public static final IntegerProperty LEAVES = BlockStateProperties.PICKLES;
+    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
+    protected static final VoxelShape ONE_SHAPE = Block.box(0, 0, 0, 16, 2, 16);
+    protected static final VoxelShape TWO_SHAPE = Block.box(0, 0, 0, 16, 3, 16);
+    protected static final VoxelShape THREE_SHAPE = Block.box(0, 0, 0, 16, 5, 16);
+
+    public DriedLeavesBlock(BlockBehaviour.Properties properties) {
+        super(properties);
+        this.registerDefaultState(this.stateDefinition.any().setValue(LEAVES, Integer.valueOf(1)).setValue(WATERLOGGED, Boolean.valueOf(true)));
+    }
+
+    @Nullable
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        BlockState blockstate = context.getLevel().getBlockState(context.getClickedPos());
+        if (blockstate.is(this)) {
+            return blockstate.setValue(LEAVES, Integer.valueOf(Math.min(3, blockstate.getValue(LEAVES) + 1)));
+        } else {
+            FluidState fluidstate = context.getLevel().getFluidState(context.getClickedPos());
+            boolean flag = fluidstate.getType() == Fluids.WATER;
+            return super.getStateForPlacement(context).setValue(WATERLOGGED, Boolean.valueOf(flag));
+        }
+    }
+
+    @Override
+    public boolean useShapeForLightOcclusion(BlockState state) {
+        return true;
+    }
+
+    protected boolean mayPlaceOn(BlockState state, LevelReader worldIn, BlockPos pos) {
+        return !state.getCollisionShape(worldIn, pos).getFaceShape(Direction.UP).isEmpty() || state.isFaceSturdy(worldIn, pos, Direction.UP);
+    }
+
+    public boolean canSurvive(BlockState state, LevelReader worldIn, BlockPos pos) {
+        BlockPos blockpos = pos.below();
+        return this.mayPlaceOn(worldIn.getBlockState(blockpos), worldIn, blockpos);
+    }
+
+    public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState, LevelAccessor worldIn, BlockPos currentPos, BlockPos facingPos) {
+        if (!stateIn.canSurvive(worldIn, currentPos)) {
+            return Blocks.AIR.defaultBlockState();
+        } else {
+            if (stateIn.getValue(WATERLOGGED)) {
+            	worldIn.scheduleTick(currentPos, Fluids.WATER, Fluids.WATER.getTickDelay(worldIn));
+            }
+
+            return super.updateShape(stateIn, facing, facingState, worldIn, currentPos, facingPos);
+        }
+    }
+
+    @SuppressWarnings("deprecation")
+	public boolean canBeReplaced(BlockState state, BlockPlaceContext useContext) {
+        return useContext.getItemInHand().getItem() == this.asItem() && state.getValue(LEAVES) < 3 || super.canBeReplaced(state, useContext);
+    }
+
+    public VoxelShape getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context) {
+        switch (state.getValue(LEAVES)) {
+            case 1:
+            default:
+                return ONE_SHAPE;
+            case 2:
+                return TWO_SHAPE;
+            case 3:
+                return THREE_SHAPE;
+        }
+    }
+
+    @SuppressWarnings("deprecation")
+	public FluidState getFluidState(BlockState state) {
+        return state.getValue(WATERLOGGED) ? Fluids.WATER.getSource(false) : super.getFluidState(state);
+    }
+
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        builder.add(LEAVES, WATERLOGGED);
+    }
+}
