+++ src/main/java/teamfusion/bettas/client/model/BettaFishModel.java
@@ -1,4 +1,4 @@
-package teamdraco.bettas.client.model;
+package teamfusion.bettas.client.model;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
