+++ src/main/java/teamfusion/bettas/init/BettasEntities.java
@@ -1,16 +1,16 @@
-package teamdraco.bettas.init;
+package teamfusion.bettas.init;
 
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
-import net.minecraftforge.registries.RegistryObject;
-import teamdraco.bettas.Bettas;
-import teamdraco.bettas.entity.BettaFishEntity;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+import teamfusion.bettas.Bettas;
+import teamfusion.bettas.entity.BettaFishEntity;
 
 public class BettasEntities {
-    public static final DeferredRegister<EntityType<?>> ENTITIES = DeferredRegister.create(ForgeRegistries.ENTITIES, Bettas.MOD_ID);
+    public static final DeferredRegister<EntityType<?>> ENTITIES = DeferredRegister.create(ForgeRegistries.ENTITY_TYPES, Bettas.MOD_ID);
 
     public static final RegistryObject<EntityType<BettaFishEntity>> BETTA_FISH = create("betta_fish", EntityType.Builder.of(BettaFishEntity::new, MobCategory.WATER_AMBIENT).sized(0.3f, 0.3f));
