+++ src/main/java/teamfusion/bettas/init/BettasFeatures.java
@@ -1,12 +1,12 @@
-package teamdraco.bettas.init;
+package teamfusion.bettas.init;
 
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.configurations.CountConfiguration;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
-import teamdraco.bettas.Bettas;
-import teamdraco.bettas.feature.MossBallFeature;
+import teamfusion.bettas.Bettas;
+import teamfusion.bettas.feature.MossBallFeature;
 
 public class BettasFeatures {
 	public static final DeferredRegister<Feature<?>> FEATURES = DeferredRegister.create(ForgeRegistries.FEATURES, Bettas.MOD_ID);
