+++ src/main/java/teamfusion/bettas/init/BettasItems.java
@@ -1,14 +1,17 @@
-package teamdraco.bettas.init;
+package teamfusion.bettas.init;
 
 import net.minecraft.world.food.FoodProperties;
-import net.minecraft.world.item.*;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.material.Fluids;
 import net.minecraftforge.common.ForgeSpawnEggItem;
-import net.minecraftforge.registries.RegistryObject;
-import teamdraco.bettas.Bettas;
-import teamdraco.bettas.item.BettasBucketItem;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+import teamfusion.bettas.Bettas;
+import teamfusion.bettas.item.BettasBucketItem;
 
 public class BettasItems {
     public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, Bettas.MOD_ID);
