+++ src/main/resources/META-INF/mods.toml
@@ -1,10 +1,10 @@
-modLoader="javafml" #mandatory
-loaderVersion="[40,)" #mandatory This is typically bumped every Minecraft version by Forge. See our download page for lists of versions.
-license="All rights reserved"
+modLoader = "javafml" #mandatory
+loaderVersion = "[43,)" #mandatory This is typically bumped every Minecraft version by Forge. See our download page for lists of versions.
+license = "All rights reserved"
 [[mods]] #mandatory
-modId="bettas" #mandatory
-version="1.0.0" #mandatory
-displayName="Bettas" #mandatory
-credits="Tazz, Ninni" #optional
-authors="Coda, superlord9362" #optional
-description='''A small mod containing a new fish: the betta! Bettas can be bucketed just like normal fish and will fight and kill eachother.'''
+modId = "bettas" #mandatory
+version = "1.0.0" #mandatory
+displayName = "Bettas" #mandatory
+credits = "Tazz, Ninni" #optional
+authors = "Coda, superlord9362, bagu_chan, Team Fusion" #optional
+description = '''A small mod containing a new fish: the betta! Bettas can be bucketed just like normal fish and will fight and kill eachother.'''
