+++ src/main/resources/data/bettas/forge/biome_modifier/betta_biome_spawn.json
@@ -1,0 +1,3 @@
+{
+  "type": "bettas:betta_biome_modifier"
+}
